Refactoring the ops from Tf1 frontend
